requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
discord==2.1.0
discord.py==2.1.0
english-words==1.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
pdf2image==1.16.2
Pillow==9.4.0
pycodestyle==2.10.0
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.13
yarl==1.8.2
